Fix niggling bug
This has been niggling me for a while, resulting in a call to the API with no module name, when looping over entries that are fetched via a relationship. Finally got to the bottom of it. It should make no functional difference to operation, but will potentially reduce the API network traffic slightly in some circumstances.